refactor: update package structure to follow onion architecture
