+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/infra/UsecaseSpringConfiguration.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright © 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.apim.infra;
+
+import io.gravitee.apim.core.application.crud_service.ApplicationCrudService;
+import io.gravitee.apim.core.log.crud_service.LogCrudService;
+import io.gravitee.apim.core.log.usecase.SearchConnectionLogUsecase;
+import io.gravitee.apim.core.plan.crud_service.PlanCrudService;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class UsecaseSpringConfiguration {
+
+    @Bean
+    public SearchConnectionLogUsecase searchConnectionLogUsecase(
+        LogCrudService logCrudService,
+        PlanCrudService planCrudService,
+        ApplicationCrudService applicationCrudService
+    ) {
+        return new SearchConnectionLogUsecase(logCrudService, planCrudService, applicationCrudService);
+    }
+}
